startup version check & fail
